feat(components): renamed input folder
